requirements.txt
# Mac写真ライブラリ移行ツール - 必要なパッケージ

osxphotos>=0.60.0
piexif>=1.1.3
Pillow>=10.0.0